requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
branca==0.5.0
certifi==2022.9.14
cffi==1.15.1
cftime==1.6.2
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
crc32c==2.3
cryptography==38.0.1
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-forms-bootstrap==3.1.0
django-views==1.0.1
exceptiongroup==1.0.0rc9
folium==0.12.1.post1
greenlet==1.1.3
gs==1.1.1
h11==0.14.0
haversine==2.7.0
html2image==2.0.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
mysqlclient==2.1.1
netCDF4==1.6.1
numpy==1.23.3
outcome==1.2.0
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pycparser==2.21
pygmt==0.7.0
PyJWT==2.5.0
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlparse==0.4.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tweak==1.0.4
tzdata==2022.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
xarray==2022.9.0
